+++ src/config.rs
@@ -1,21 +1,17 @@
 use anyhow::{anyhow, Context, Result};
-use clap::Args;
 use figment::{
     providers::{Env, Serialized},
     Figment,
 };
 use itertools::iproduct;
 use once_cell::sync::Lazy;
-use serde::{Deserialize, Serialize};
-use serde_with::{
-    formats::{CommaSeparator, Separator},
-    serde_as, skip_serializing_none, BoolFromInt, PickFirst, StringWithSeparator,
-};
 use std::{
     env, fs,
     io::{Error, ErrorKind},
     path::{Path, PathBuf},
 };
+
+use crate::Flags;
 
 static COMPOSE_FILE_NAMES: Lazy<Vec<String>> = Lazy::new(|| {
     iproduct!(["compose", "docker-compose"], ["yaml", "yml"])
@@ -23,63 +19,14 @@ static COMPOSE_FILE_NAMES: Lazy<Vec<String>> = Lazy::new(|| {
         .collect()
 });
 
-struct PathSeparator;
-
-impl Separator for PathSeparator {
-    fn separator() -> &'static str {
-        Box::leak(
-            env::var("COMPOSE_PATH_SEPARATOR")
-                .unwrap_or_else(|_| {
-                    String::from(if cfg!(unix) {
-                        ":"
-                    } else if cfg!(windows) {
-                        ";"
-                    } else {
-                        unreachable!()
-                    })
-                })
-                .into_boxed_str(),
-        )
-    }
-}
-
-#[skip_serializing_none]
-#[serde_as]
-#[derive(Args, Serialize, Deserialize, Debug)]
+#[derive(Debug)]
 pub(crate) struct Config {
-    /// Project name
-    #[arg(short, long)]
     pub(crate) project_name: Option<String>,
-
-    /// Compose configuration files
-    #[arg(short, long)]
-    #[serde_as(as = "Option<PickFirst<(_, StringWithSeparator::<PathSeparator, String>)>>")]
-    pub(crate) file: Option<Vec<String>>,
-
-    /// Specify a profile to enable
-    #[arg(long)]
-    #[serde_as(as = "Option<PickFirst<(_, StringWithSeparator::<CommaSeparator, String>)>>")]
-    #[serde(rename = "profiles")]
-    pub(crate) profile: Option<Vec<String>>,
-
-    /// Specify an alternate environment file
-    #[arg(long)]
-    pub(crate) env_file: Option<String>,
-
-    /// Specify an alternate working directory
-    #[arg(long)]
-    pub(crate) project_directory: Option<String>,
-
-    #[arg(skip)]
-    #[serde_as(as = "Option<PickFirst<(_, BoolFromInt)>>")]
-    pub(crate) convert_windows_paths: Option<bool>,
-
-    #[arg(skip)]
-    pub(crate) path_separator: Option<String>,
-
-    #[arg(skip)]
-    #[serde_as(as = "Option<PickFirst<(_, BoolFromInt)>>")]
-    pub(crate) ignore_orphans: Option<bool>,
+    pub(crate) files: Vec<String>,
+    pub(crate) profiles: Vec<String>,
+    pub(crate) project_directory: String,
+    pub(crate) convert_windows_paths: bool,
+    pub(crate) ignore_orphans: bool,
 }
 
 fn find(directory: &Path, files: &Vec<String>) -> Result<PathBuf> {
@@ -104,15 +51,24 @@ fn find(directory: &Path, files: &Vec<String>) -> Result<PathBuf> {
     }
 }
 
-fn resolve(config: &Config) -> Result<Config> {
-    let mut config = Figment::new()
+fn resolve(flags: &Flags) -> Result<Config> {
+    let flags = Figment::new()
         .merge(Env::prefixed("COMPOSE_").ignore(&["env_file", "project_directory"]))
-        .merge(Serialized::defaults(config))
-        .extract::<Config>()?;
+        .merge(Serialized::defaults(flags))
+        .extract::<Flags>()?;
 
-    if config.file.is_none() {
+    let files = if let Some(files) = flags.file {
+        files
+            .iter()
+            .map(|file| {
+                fs::canonicalize(file)
+                    .with_context(|| anyhow!("{file} not found"))
+                    .map(|file| file.to_string_lossy().to_string())
+            })
+            .collect::<Result<Vec<_>, _>>()?
+    } else {
         let file = find(
-            &config
+            &flags
                 .project_directory
                 .as_ref()
                 .map_or(env::current_dir()?, PathBuf::from),
@@ -124,51 +80,54 @@ fn resolve(config: &Config) -> Result<Config> {
             file.extension().unwrap().to_string_lossy()
         ));
 
-        config.file = Some(
-            if override_file.is_file() {
-                vec![&file, &override_file]
-            } else {
-                vec![&file]
-            }
-            .into_iter()
-            .map(|file| file.to_string_lossy().to_string())
-            .collect(),
-        );
-    }
-
-    if let Some(files) = &mut config.file {
-        for file in files {
-            *file = fs::canonicalize(&file)
-                .with_context(|| anyhow!("{file} not found"))?
-                .to_string_lossy()
-                .to_string();
+        if override_file.is_file() {
+            vec![&file, &override_file]
+        } else {
+            vec![&file]
         }
-    }
+        .into_iter()
+        .map(|file| {
+            file.canonicalize()
+                .with_context(|| anyhow!("{} not found", file.to_string_lossy()))
+                .map(|file| file.to_string_lossy().to_string())
+        })
+        .collect::<Result<Vec<_>, _>>()?
+    };
 
-    if let Some(file) = config.file.as_ref().and_then(|files| files.first()) {
-        let project_directory = config.project_directory.get_or_insert_with(|| {
-            Path::new(file)
-                .parent()
-                .map(|parent| parent.to_string_lossy().to_string())
-                .unwrap_or_default()
-        });
-
-        *project_directory = fs::canonicalize(&project_directory)
-            .with_context(|| anyhow!("{project_directory} not found"))?
+    let project_directory = if let Some(dir) = &flags.project_directory {
+        fs::canonicalize(dir)
+            .with_context(|| anyhow!("{dir} not found"))?
             .to_string_lossy()
-            .to_string();
-    }
+            .to_string()
+    } else {
+        let parent = Path::new(&files[0])
+            .parent()
+            .unwrap_or_else(|| Path::new("/"));
 
-    Ok(config)
+        parent
+            .canonicalize()
+            .with_context(|| anyhow!("{} not found", parent.to_string_lossy()))?
+            .to_string_lossy()
+            .to_string()
+    };
+
+    Ok(Config {
+        project_name: flags.project_name,
+        files,
+        profiles: flags.profile.unwrap_or_default(),
+        project_directory,
+        convert_windows_paths: flags.convert_windows_paths.unwrap_or_default(),
+        ignore_orphans: flags.ignore_orphans.unwrap_or_default(),
+    })
 }
 
-pub(crate) fn load(config: &Config) -> Result<Config> {
-    let resolved_config = resolve(config)?;
-    let env_file = config.env_file.as_ref().map_or_else(
-        || PathBuf::from(resolved_config.project_directory.unwrap()).join(".env"),
+pub(crate) fn load(flags: Flags) -> Result<Config> {
+    let config = resolve(&flags)?;
+    let env_file = flags.env_file.as_ref().map_or_else(
+        || PathBuf::from(config.project_directory).join(".env"),
         PathBuf::from,
     );
 
     dotenvy::from_filename(env_file).ok();
-    resolve(config)
+    resolve(&flags)
 }